more refactor and added example to for types usage
