+++ packages/cli/src/lib/setup/setupSetup.ts
@@ -30,7 +30,7 @@ import * as pluginInfos from '@/lib/setup/pluginInfos.js';
 import rl from 'readline-sync';
 import { FORBIDDEN_CHARS, getHostObject } from '@iobroker/js-controller-common-db/tools';
 import os from 'node:os';
-import { SYSTEM_ADAPTER_PREFIX, SYSTEM_HOST_PREFIX } from '@iobroker/js-controller-common/constants';
+import { SYSTEM_ADAPTER_PREFIX, SYSTEM_HOST_PREFIX } from '@iobroker/js-controller-common-db/constants';
 import { Upload } from '@/lib/setup/setupUpload.js';
 import { createRequire } from 'node:module';
 import * as url from 'node:url';
