+++ packages/common-db/src/lib/common/constants.ts
