+++ packages/common-db/src/lib/common/tools.ts
@@ -3538,70 +3538,6 @@ export function getLogger(log: any): InternalLogger {
     return log;
 }
 
-export interface OrderedInstancesObject {
-    1: ioBroker.InstanceObject[];
-    2: ioBroker.InstanceObject[];
-    3: ioBroker.InstanceObject[];
-    admin: ioBroker.InstanceObject[];
-}
-
-/**
- * Get ordered instances according to tier level
- *
- * @param objects - Objects DB
- * @param logger - logger object
- * @param logPrefix prefix for logging
- */
-export async function getInstancesOrderedByStartPrio(
-    objects: any,
-    logger: any,
-    logPrefix = ''
-): Promise<ioBroker.InstanceObject[]> {
-    const instances: OrderedInstancesObject = { 1: [], 2: [], 3: [], admin: [] };
-    const allowedTiers = [1, 2, 3];
-
-    if (logPrefix) {
-        // append space if we have a prefix
-        logPrefix += ' ';
-    }
-
-    let doc: GetObjectViewResult<ioBroker.InstanceObject> = { rows: [] };
-    try {
-        doc = await objects.getObjectViewAsync('system', 'instance', {
-            startkey: 'system.adapter.',
-            endkey: 'system.adapter.\u9999'
-        });
-    } catch (e) {
-        if (e.message?.startsWith('Cannot find ')) {
-            logger.error(`${logPrefix} _design/system missing - call node ${appName}.js setup`);
-        } else {
-            logger.error(`${logPrefix} Can not get instances: ${e.message}`);
-        }
-    }
-
-    if (!doc.rows || doc.rows.length === 0) {
-        logger.info(`${logPrefix} no instances found`);
-    } else {
-        for (const row of doc.rows) {
-            if (row?.value) {
-                if (row.value._id.startsWith('system.adapter.admin')) {
-                    instances.admin.push(row.value);
-                } else if (
-                    row.value.common?.tier !== undefined &&
-                    allowedTiers.includes(Math.round(row.value.common.tier))
-                ) {
-                    instances[row.value.common.tier].push(row.value);
-                } else {
-                    // no valid tier, so put it in the last one
-                    instances['3'].push(row.value);
-                }
-            }
-        }
-    }
-
-    return [...instances.admin, ...instances['1'], ...instances['2'], ...instances['3']];
-}
-
 /**
  * Set capabilities of the given executable on Linux systems
  *
